+++ src/pages/AllowedServicePage/BoxEachUrlInfo/BoxEachUrlInfo.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 
-import { UrlInfo } from './../../types';
+import { UrlInfo } from '../types';
 
 interface BoxEachUrlInfoProps {
 	urlInfo: UrlInfo;
