+++ src/pages/AllowedServicePage/ColorPallete/ColorPallete.tsx
