+++ src/pages/AllowedServicePage/InputAllowedServiceUrl/InputAllowedServiceUrl.tsx
@@ -5,9 +5,9 @@ import { isUrlValid } from '@/shared/utils/isUrlValid/index';
 import InputClearButton from '@/shared/assets/svgs/btn_inputClear.svg?react';
 import IconInputError from '@/shared/assets/svgs/error_input.svg?react';
 
-import ButtonAddAllowedService from '@/pages/MoribSetPage/components/Button/ButtonAddAllowedService';
+import ButtonAddAllowedService from '../ButtonAddAllowedService/ButtonAddAllowedService';
 
-import { UrlInfo } from './../../types';
+import { UrlInfo } from '../types';
 
 interface InputAllowedServiceUrlProps {
 	urlList: UrlInfo[];
